gcp/functions/scraper/requirements.txt
functions-framework==3.*
google-cloud-firestore
google-cloud-secret-manager
selenium==4.15.0
requests
flask

